python_code: Eggsecutables need a top-level main
Executable eggs need a top-level __main__.py to respond to the
interpreter.  I thought I could just read down inside the package and
use the package's __main__.py (idk, might delete later) but that wasn't
working.

I transitioned to adding a __main__.py at the level of the script that
calls setup() and adding the __main__ as a py_module (keywork argument)
in the call to setup().

Silly me.  What was I thinking?!

Changes to be committed:
	modified:   src/python/python_package/__main__.py
	modified:   src/python/setup_python_script05.py