python_code: Extricated some functionality
Extricated the prepare_file_output functionality into it's own module to
avoid duplicate code.

NOTE: If I'm not using it everywhere then it's because there are some
python_script examples (e.g., python_script01.py) that should be
entirely self-contained.  The functional test I designed requires
it/them to be so.

Changes to be committed:
	new file:   src/python/prepare_data.py
	modified:   src/python/python_module.py