system: run_system.sh Incremental Commit
This is a step on the path to automating, through indirection, how each
of the six Python examples are executed.  This script may fail for
number 5 but that's ok because it's a directory, not a file.  I was
really looking to automate the list of commands python-style.

Changes to be committed:
	modified:   devops/scripts/run_system.sh