popen: main_popen.c Inital Commit
Even though this is the initial commit, this code appears to be entirely
functional.  I functionally tested it... manually.  It appears to work
as expected.

The next step is to write the shell script that utilizes the popen.bin
binary to execute all six Python code examples.  Then, integrate that
new script into a "run all" shell script to aid in ease of use.

Changes to be committed:
	new file:   src/c/main_popen.c