python_linking: Linux Journal code Initial Commit
Initial commit of the code from:
	https://www.linuxjournal.com/article/8497

SPOILER ALERT: This code will not compile.
	1. There are typos (I'll fix and commit later)
	2. I don't have Python2.3 locally and I don't care to roll back.
I intend to try and make this code work with Python2.7, then Python3.
	3. Never number a list unless you have at least three items

NOTE: I am using this commit to document(?) the version of the code I
found on the Linux Journal website.

Changes to be committed:
	new file:   src/c/linux_journal_code.c