python_linking: Fixed libpython.c BUG
Fixed a memory leak.
I was free()ing the contents of the array but I wasn't free()ing the
array pointer itself.
FIXED.

NOTE: valgrind didn't come up with this (because valgrind doesn't like
the way libpython does business?).  I came up with this during a code
review.  I'm tempted to put memwatch in there...

Changes to be committed:
	modified:   src/c/main_libpython.c